apps/meetmate/requirements/base.txt
channels
channels_redis
Celery
Django==4.2.6
djangorestframework
django-allauth
django-redis
django-quill-editor
django-debug-toolbar

google-api-python-client==2.104.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.3
httplib2==0.22.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
requests==2.31.0

ics
paypal-checkout-serversdk
pdfkit
Pillow==8.3.1
psycopg2
pytz
redis
rsa
simple-crypt
stripe>=2.48.0,<3.0